move registry to it's own file and fix activePlugins updating based on ctx
